Remove leader meeting reminders over break, fix some return annotation errors
