+++ src/constants.py
@@ -1,0 +1,7 @@
+import datetime
+
+SEMESTERS = [
+    (datetime.date(2023, 8, 23), datetime.date(2023, 12, 6)),
+    (datetime.date(2024, 1, 8), datetime.date(2024, 4, 24)),
+    (datetime.date(2024, 5, 13), datetime.date(2024, 8, 9)),
+]
